+++ pages/login.tsx
@@ -1,11 +1,5 @@
-import MainContainer from "@src/components/MainContainer";
-
 const Login = () => {
-  return (
-    <MainContainer>
-      <h1>Login</h1>
-    </MainContainer>
-  );
+  return <h1>Login</h1>;
 };
 
 export default Login;
