+++ pages/profile.tsx
@@ -1,11 +1,5 @@
-import MainContainer from "@src/components/MainContainer";
-
 const Profile = () => {
-  return (
-    <MainContainer>
-      <h1>Profile</h1>
-    </MainContainer>
-  );
+  return <h1>Profile</h1>;
 };
 
 export default Profile;
