+++ pages/register.tsx
@@ -1,11 +1,5 @@
-import MainContainer from "@src/components/MainContainer";
-
 const Register = () => {
-  return (
-    <MainContainer>
-      <h1>Register</h1>
-    </MainContainer>
-  );
+  return <h1>Register</h1>;
 };
 
 export default Register;
