+++ src/components/Header/components/DesktopMenu/DesktopMenu.tsx
@@ -1,0 +1,26 @@
+import Link from "next/link";
+import React from "react";
+import { Box, Button, Typography } from "@mui/material";
+import { PAGES } from "../../Header";
+
+function DesktopMenu() {
+  return (
+    <Box>
+      <Box display={{ sm: "none", md: "block" }}>
+        {PAGES.map((page) => (
+          <Button key={page.name}>
+            <Link href={page.href} passHref>
+              <a>
+                <Typography variant="h5" component="h2">
+                  {page.name}
+                </Typography>
+              </a>
+            </Link>
+          </Button>
+        ))}
+      </Box>
+    </Box>
+  );
+}
+
+export default DesktopMenu;
