+++ src/components/Header/components/DesktopMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DesktopMenu";
