+++ src/components/Header/components/HeaderLogo/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HeaderLogo";
