+++ src/components/Layout/LayoutWrapper.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import Header from "../Header";
+import MainContainer from "../MainContainer";
 import * as Styles from "./styles";
 
 export type LayoutProps = { children?: React.ReactNode | HTMLElement };
@@ -8,7 +9,7 @@ function LayoutWrapper({ children }: LayoutProps) {
   return (
     <Styles.Element>
       <Header />
-      {children}
+      <MainContainer>{children}</MainContainer>
     </Styles.Element>
   );
 }
